Move the setloading text action to the App user action module
