+++ src/Kiroku.tsx
@@ -14,6 +14,7 @@ import {useFirebase} from '@context/global/FirebaseContext';
 import {useUserConnection} from '@context/global/UserConnectionContext';
 import SplashScreenStateContext from '@context/global/SplashScreenStateContext';
 import {useConfig} from '@context/global/ConfigContext';
+import * as User from './libs/actions/User';
 import Navigation from './libs/Navigation/Navigation';
 import NavigationRoot from './libs/Navigation/NavigationRoot';
 // import PushNotification from '@libs/Notification/PushNotification';
@@ -23,7 +24,6 @@ import SplashScreenHider from './components/SplashScreenHider';
 import * as ActiveClientManager from './libs/ActiveClientManager';
 import * as ErrorUtils from './libs/ErrorUtils';
 import * as UserUtils from './libs/UserUtils';
-import * as User from '@userActions/User';
 // import StartupTimer from '@libs/StartupTimer';
 import Visibility from './libs/Visibility';
 import ONYXKEYS from './ONYXKEYS';
@@ -37,7 +37,7 @@ import UserOfflineModal from './components/UserOfflineModal';
 import CONFIG from './CONFIG';
 import UpdateAppModal from './components/UpdateAppModal';
 import VerifyEmailModal from './components/VerifyEmailModal';
-import AgreeToTermsModal from '@components/AgreeToTermsModal';
+import AgreeToTermsModal from './components/AgreeToTermsModal';
 import FullScreenLoadingIndicator from './components/FullscreenLoadingIndicator';
 import CONST from './CONST';
 import ERRORS from './ERRORS';
